Refactor of new member duplicate email detection
Prior to this change attempting to add a member without an email address
would result in a search page containing every single member in the
database without an email address, which is suboptimal.

(Not allowing members with no email address was why a test was failing.)

Also change to redirect to the search results URL, with the correct
email query set, rather than returning the search results (with edit
box) directly in the 200 response, to avoid reload/history issues.